various components/shared refactoring.
