+++ app/components/shared/LinkToAccounts/LinkToAccounts.jsx
@@ -1,12 +1,11 @@
-// @flow
 import { Tooltip } from "pi-ui";
 import { Link } from "react-router-dom";
 import { FormattedMessage as T } from "react-intl";
-import "style/MiscComponents.less";
+import styles from "./LinkToAccounts.module.css";
 
 const LinkToAccounts = () => (
   <Tooltip content={<T id="accountsButton.tip" m="Accounts" />}>
-    <Link to="/accounts" className="accounts-button-icon" />
+    <Link to="/accounts" className={styles.icon} />
   </Tooltip>
 );
