+++ app/components/shared/LinkToAccounts/LinkToAccounts.module.css
@@ -1,0 +1,19 @@
+.icon {
+  display: inline-block;
+  width: 20px;
+  height: 20px;
+  background-image: var(--accounts-button-icon);
+  background-position: 50%;
+  background-size: 12px auto;
+  background-repeat: no-repeat;
+  box-shadow: 0 3px 8px var(--input-color-shadow);
+  border-radius: 5px;
+  background-color: var(--input-color);
+  cursor: pointer;
+  vertical-align: middle;
+}
+
+.icon:hover {
+  box-shadow: 0 1px 3px var(--input-color-shadow);
+  background-color: var(--account-button-hover);
+}
