+++ app/components/shared/LinkToAccounts/index.js
@@ -1,0 +1,1 @@
+export { default } from "./LinkToAccounts";
