--- app/components/shared/ShowError.js
@@ -1,6 +1,0 @@
-const ShowError = ({ error, className }) =>
-  !error ? null : (
-    <div className={"error" + (className ? " " + className : "")}>{error}</div>
-  );
-
-export default ShowError;
