+++ app/components/shared/ShowError.jsx
@@ -1,0 +1,6 @@
+import { classNames } from "pi-ui";
+
+const ShowError = ({ error, className }) =>
+  !error ? null : <div className={classNames("error", className)}>{error}</div>;
+
+export default ShowError;
