+++ app/components/shared/ShowWarning.jsx
@@ -1,5 +1,7 @@
+import { classNames } from "pi-ui";
+
 const ShowWarning = ({ warn, className }) => (
-  <div className={["warning", className].join(" ")}>{warn}</div>
+  <div className={classNames("warning", className)}>{warn}</div>
 );
 
 ShowWarning.propTypes = {
