+++ app/components/shared/StaticSwitch.jsx
