+++ app/components/shared/UnsignedTx/UnsignedTx.jsx
@@ -1,0 +1,15 @@
+import { CopyToClipboard } from "shared";
+import styles from "./UnsignedTx.module.css";
+
+const UnsignedTx = ({ tx, title }) => (
+  <div className={styles.txArea}>
+    <div className={styles.txTitle}>{title}</div>
+    <div className={styles.tx}>{tx}</div>
+    <CopyToClipboard
+      textToCopy={tx}
+      className={styles.copyIcon}
+    />
+  </div>
+);
+
+export default UnsignedTx;
