+++ app/components/shared/UnsignedTx/index.js
@@ -1,0 +1,1 @@
+export { default } from "./UnsignedTx";
