+++ app/connectors/index.js
@@ -3,5 +3,4 @@ export { default as send } from "./send";
 export { default as ticketsList } from "./ticketsList";
 export { default as ticketsOverview } from "./ticketsOverview";
 export { default as routing } from "./routing";
-export { default as modalVisible } from "./modalVisible";
 export { default as lnPage } from "./lnPage";
