+++ app/hooks/index.js
@@ -9,3 +9,4 @@ export { default as useClickOutside } from "./useClickOutside";
 export { default as useTheming } from "./useTheming";
 export { default as useTrezor } from "./useTrezor";
 export { default as useLocale } from "./useLocale";
+export { default as useNetwork } from "./useNetwork";
